Eliminated duplicate code from Pod and Container classes
1. Deleted identical PodRecommendations.java and ContainerRecommendations.java
and consolidated their methods in RecommendationsImpl.java
2. Renamed AbstractRecommendations.java to RecommendationsImpl.java
3. Deleted KubernetesApplicationRecommendations and DockerApplicationRecommendations.java
and moved their common methods to ApplicationRecommendationsImpl. This also results
in Docker now suggesting resource requests.
4. Removed identical classes KubernetesAnalysisImpl and DockerAnalysisImpl
