+++ src/main/java/com/kruize/environment/EnvTypeImpl.java
@@ -21,7 +21,7 @@ import com.kruize.environment.docker.DockerEnvImpl;
 import com.kruize.environment.kubernetes.KubernetesEnvImpl;
 import com.kruize.metrics.AbstractMetrics;
 import com.kruize.query.Query;
-import com.kruize.recommendations.application.AbstractApplicationRecommendations;
+import com.kruize.recommendations.application.ApplicationRecommendationsImpl;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -31,7 +31,7 @@ public abstract class EnvTypeImpl implements EnvType
     public AbstractMetrics metrics;
     public Analysis analysis;
     public Query query;
-    public AbstractApplicationRecommendations applicationRecommendations;
+    public ApplicationRecommendationsImpl applicationRecommendations;
 
     private static EnvTypeImpl envType = null;
